backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==1.4.47
pydantic==1.10.7
python-dotenv==1.0.0
pymysql==1.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.0.0
boto3==1.26.115
cryptography==40.0.1
bcrypt==4.0.1